+++ python/core/auto_generated/qgsapplication.sip.in
@@ -823,6 +823,13 @@ Returns the application's bookmark manager, used for storing installation-wide b
 %End
 
 
+    static QgsRecentStyleHandler *recentStyleHandler() /KeepReference/;
+%Docstring
+Returns the handler for recently used style items.
+
+.. versionadded:: 3.22
+%End
+
     static QgsStyleModel *defaultStyleModel();
 %Docstring
 Returns a shared :py:class:`QgsStyleModel` containing the default style library (see :py:func:`QgsStyle.defaultStyle()`).
