+++ python/gui/auto_generated/qgsgui.sip.in
@@ -153,13 +153,6 @@ Returns the global relation widget registry, used for managing all known relatio
 .. versionadded:: 3.18
 %End
 
-    static QgsRecentStyleHandler *recentStyleHandler() /KeepReference/;
-%Docstring
-Returns the handler for recently used style items.
-
-.. versionadded:: 3.22
-%End
-
     static void enableAutoGeometryRestore( QWidget *widget, const QString &key = QString() );
 %Docstring
 Register the widget to allow its position to be automatically saved and restored when open and closed.
