+++ src/core/qgsrecentstylehandler.cpp
